fix warnings on windows x64 builds src and mods projects - only libsofia included on the libs side

+++ src/include/switch_types.h
@@ -770,9 +770,9 @@ typedef struct {
 	const char *T38FaxUdpEC;
 	const char *T38VendorInfo;
 	const char *remote_ip;
-	uint32_t remote_port;
+	uint16_t remote_port;
 	const char *local_ip;
-	uint32_t local_port;
+	uint16_t local_port;
 } switch_t38_options_t;
 
 /*!
